+++ src/components/signals/SignalsList.tsx
@@ -1,19 +1,12 @@
-import {
-  CompassOutlined,
-  EnvironmentOutlined,
-  ReloadOutlined,
-} from "@ant-design/icons";
-import { Card, Col, List, Row, Typography, Button } from "antd";
-import dayjs from "dayjs";
+import { ReloadOutlined } from "@ant-design/icons";
+import { Card, Col, List, Row, Typography } from "antd";
 import _ from "lodash";
 import React, { useContext, useEffect, useMemo, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import config from "../../config.json";
 import WhiteFlagContext from "../../helpers/Context";
 import { useApi } from "../../hooks/useApi";
 import { DecodedSignal } from "../../models/DecodedSignal";
 import {
-  InfrastructureSubjectCode,
   WhiteflagResponse,
   WhiteflagSignal,
 } from "../../models/WhiteflagSignal";
@@ -23,18 +16,18 @@ import CoordinatesHeader from "../layout/CoordinatesHeader";
 import PageToggle from "../layout/PageToggle";
 import { AddSignalDrawer } from "./AddSignalDrawer";
 import { SignalDetailDrawer } from "./SignalDetailDrawer";
-import { SignalBodyText } from "../../models/SignalBodyText";
-import { formatCoordinate } from "../../helpers/CoordinatesHelper";
+import SignalCard from "./SignalCard";
 export interface Location {
   latitude?: number;
   longitude?: number;
 }
 
 export const SignalsList: React.FC = () => {
-  const navigate = useNavigate();
   const ctx = useContext(WhiteFlagContext);
-  const [locationModalVisable, setLocationModalVisable] = useState<boolean>(false);
-  const [newSignalDrawerOpen, setNewSignalDrawerOpen] = useState<boolean>(false);
+  const [locationModalVisable, setLocationModalVisable] =
+    useState<boolean>(false);
+  const [newSignalDrawerOpen, setNewSignalDrawerOpen] =
+    useState<boolean>(false);
   const [activeSignal, setActiveSignal] = useState<DecodedSignal>();
   const [distanceToSignal, setDistanceToSignal] = useState<number>();
   const [isLoading, setIsLoading] = useState<boolean>(true);
@@ -42,7 +35,7 @@ export const SignalsList: React.FC = () => {
     entities: signalResponses,
     endpoints: signalsEndpoint,
     loading: isLoadingSignals,
-    error: signalsError
+    error: signalsError,
   } = useApi<WhiteflagSignal, WhiteflagResponse>({
     url: `${config.baseUrl}${Settings.endpoints.signals.get}`,
   });
@@ -50,7 +43,7 @@ export const SignalsList: React.FC = () => {
   const {
     endpoints: decodeListEndpoint,
     loading: isLoadingDecodeList,
-    error: decodeListError
+    error: decodeListError,
   } = useApi<WhiteflagResponse[]>({
     url: `${config.baseUrl}${Settings.endpoints.whiteflag.decodeList}`,
   });
@@ -100,8 +93,7 @@ export const SignalsList: React.FC = () => {
     if (!ctx.whiteflagSignals) {
       setIsLoading(true);
     }
-    const ids = signalResponses
-      .map((response) => response.id);
+    const ids = signalResponses.map((response) => response.id);
     const whiteflagResponse = await decodeListEndpoint.directPost({
       signals: ids,
     });
@@ -149,20 +141,23 @@ export const SignalsList: React.FC = () => {
   };
 
   const validSignals = useMemo(() => {
-    const signalsWithValidCoordinates = ctx?.filteredWhiteflagTextSignals.filter(signal => {
-      const coordinates = ctx.extractCoordinates(signal);
-      return coordinates !== null;
-  });
+    const signalsWithValidCoordinates =
+      ctx?.filteredWhiteflagTextSignals.filter((signal) => {
+        const coordinates = ctx.extractCoordinates(signal);
+        return coordinates !== null;
+      });
 
-    const sortedSignals = signalsWithValidCoordinates.sort(ctx.compareDistances)
+    const sortedSignals = signalsWithValidCoordinates.sort(
+      ctx.compareDistances
+    );
 
-    return sortedSignals
+    return sortedSignals;
   }, [ctx?.filteredWhiteflagTextSignals, ctx.location]);
 
   useEffect(() => {
     ctx.setValidSignals(validSignals);
   }, [validSignals, ctx.setValidSignals]);
-  
+
   return (
     <React.Fragment>
       <CoordinatesHeader />
@@ -196,7 +191,7 @@ export const SignalsList: React.FC = () => {
                 display: "flex",
                 alignItems: "center",
                 color: "#FFFFFF",
-                cursor: "pointer"
+                cursor: "pointer",
               }}
               onClick={() => {
                 getAllSignals();
@@ -207,7 +202,7 @@ export const SignalsList: React.FC = () => {
                 style={{
                   color: "#FFFFFF",
                   fontSize: "16px",
-                  cursor: "pointer"
+                  cursor: "pointer",
                 }}
               />
               <span style={{ paddingLeft: "5px", fontSize: "14px" }}>
@@ -220,161 +215,18 @@ export const SignalsList: React.FC = () => {
               gutter: 16,
               xs: 1,
               sm: 2,
-              md: 4,
-              lg: 4,
-              xl: 6,
-              xxl: 3,
+              md: 3,
+              lg: 3,
+              xl: 4,
+              xxl: 5,
             }}
             loading={isLoadingSignals || isLoadingDecodeList}
             dataSource={validSignals}
-            style={{ width: "100%" }}
+            // style={{ width: "100%" }}
             renderItem={(signal) => {
-              const { latitude, longitude } = ctx.extractCoordinates(signal);
-              const bearing = ctx.calculateBearing(signal);
-              const subjectCodeIndex = Object.values(
-                InfrastructureSubjectCode
-              ).indexOf(signal.signal_body.subjectCode);
-
-              const texts = signal?.signal_body?.text
-                ? (JSON.parse(signal.signal_body.text) as SignalBodyText)
-                : undefined;
-
               return (
                 <List.Item>
-                  <Card
-                    hoverable
-                    bordered={false}
-                    bodyStyle={{ padding: "16px" }}
-                    style={{
-                      marginLeft: "16px",
-                      marginRight: "16px",
-                    }}
-                    onClick={() => ctx.handleSignalSelect(signal)}
-                  >
-                    <Row>
-                      <Typography.Text
-                        type={"secondary"}
-                        style={{ color: "#FFFFFF", fontSize: "18px" }}
-                      >
-                        {
-                          Object.keys(InfrastructureSubjectCode)[
-                            subjectCodeIndex
-                          ]
-                        }
-                      </Typography.Text>
-                    </Row>
-                    <Row>
-                      <Typography.Title
-                        level={1}
-                        style={{ fontWeight: "normal", marginTop: "0px" }}
-                      >
-                        {texts?.name}
-                      </Typography.Title>
-                    </Row>
-                    <Row style={{ display: "flex" }}>
-                      <CompassOutlined
-                        style={{
-                          paddingRight: "10px",
-                          height: "24px",
-                          width: "24px",
-                        }}
-                      />
-                      <div>
-                        <Row>
-                          <Typography.Text style={{ marginTop: "0px" }}>
-                            {bearing
-                              ? `${ctx.calculateDistanceToSignal(ctx.extractCoordinates(signal)
-                                  )?.toFixed(2)} km · ${bearing?.toFixed(
-                                  0
-                                )}° ${ctx.getCompassDirection(bearing!)}`
-                              : "Provide reference location"}
-                          </Typography.Text>
-                        </Row>
-                        <Row>
-                          <Typography.Text
-                            type={"secondary"}
-                            style={{ color: "#FFFFFF" }}
-                          >
-                            {`${latitude ? formatCoordinate('latitude', latitude): 0}, 
-                            ${longitude ? formatCoordinate('longitude', longitude): 0}`}
-                          </Typography.Text>
-                        </Row>
-                      </div>
-                    </Row>
-                    <div style={{ paddingTop: "16px" }}>
-                      <Row>
-                        <Typography.Text
-                          type={"secondary"}
-                          style={{ color: "#FFFFFF" }}
-                        >
-                          Uploaded by
-                        </Typography.Text>
-                      </Row>
-                      <Row>
-                        <Typography.Text>
-                          {signal.sender.username}
-                        </Typography.Text>
-                      </Row>
-                    </div>
-                    <div style={{ paddingTop: "16px" }}>
-                      <Row>
-                        <Typography.Text
-                          type={"secondary"}
-                          style={{ color: "#FFFFFF" }}
-                        >
-                          Uploaded on
-                        </Typography.Text>
-                      </Row>
-                      <Row>
-                        <Typography.Text>
-                          {dayjs(signal?.timestamp).format(
-                            "D MMMM YYYY, HH:mm"
-                          )}
-                        </Typography.Text>
-                      </Row>
-                      <Row>
-                        <Typography.Text>{`by ${signal.sender_group}`}</Typography.Text>
-                      </Row>
-                    </div>
-                    <Row>
-                      <Button
-                        type="default"
-                        style={{
-                          display: "block",
-                          borderRadius: "16px",
-                          fontWeight: 500,
-                          marginTop: "15px",
-                          backgroundColor: "#FFFFFF00",
-                          borderColor: "#FFFFFF",
-                          color: "#FFFFFF",
-                          marginRight: "12px",
-                        }}
-                        onClick={() => {
-                          ctx.mapNavigationHandler(Number.parseFloat(latitude).toFixed(8), Number.parseFloat(longitude).toFixed(8));
-                          ctx.activeSignalHandler(signal);
-                          navigate("/maps");
-                        }}>
-                        Show on map
-                      </Button>
-                      <Button
-                        type="default"
-                        style={{
-                          display: "block",
-                          borderRadius: "16px",
-                          fontWeight: 500,
-                          marginTop: "15px",
-                          backgroundColor: "#FFFFFF00",
-                          borderColor: "#FFFFFF",
-                          color: "#FFFFFF",
-                        }}
-                        icon={<EnvironmentOutlined />}
-                        href={`https://www.google.com/maps/dir/${ctx.location.latitude},${ctx.location.longitude}/${latitude},${longitude}`}
-                        target="_blank"
-                        >
-                          Show route
-                      </Button>
-                    </Row>
-                  </Card>
+                  <SignalCard ctx={ctx} signal={signal} />
                 </List.Item>
               );
             }}
@@ -401,7 +253,11 @@ export const SignalsList: React.FC = () => {
         </React.Fragment>
       )}
       {!newSignalDrawerOpen && (
-        <PageToggle setNewSignalDrawerOpen={setNewSignalDrawerOpen} hideFirstButton={!!activeSignal} hideSecondButton={!!activeSignal} />
+        <PageToggle
+          setNewSignalDrawerOpen={setNewSignalDrawerOpen}
+          hideFirstButton={!!activeSignal}
+          hideSecondButton={!!activeSignal}
+        />
       )}
       <AddSignalDrawer
         open={newSignalDrawerOpen}
