+++ src/styles/components/_list.scss
@@ -1,16 +1,40 @@
 .ant-spin {
   color: #1085cd;
 }
+
 .ant-spin .ant-spin-dot-item {
   background-color: #1085cd;
 }
+
 .ant-spin-nested-loading .ant-spin-container.ant-spin-blur {
   opacity: 0;
 }
-.ant-list-grid .ant-col > .ant-list-item {
-  margin-block-end: 8px;
+
+.ant-list-grid .ant-col {
+  .ant-list-item {
+    margin-block-end: 16px;
+    margin-bottom: 16px;
+    height: 100%;
+  }
 }
-.ant-list-grid .ant-col > .ant-list-item .ant-card,
-.ant-list-grid .ant-col > .ant-list-item .ant-card-body {
-  border-radius: 16px;
+
+.ant-list-grid .ant-col {
+  height: 100%;
+
+  .ant-list-item .ant-card,
+  .ant-list-item .ant-card-body {
+    border-radius: 16px;
+    height: calc(100% - 20px);
+  }
+}
+
+.signal-card__button-row {
+  width: 100%;
+
+  .ant-btn {
+    flex: 1;
+    span {
+      font-size: 14px;
+    }
+  }
 }
